+++ blade-graphics/src/traits.rs
@@ -43,11 +43,15 @@ pub trait ShaderDevice {
 
 pub trait CommandDevice {
     type CommandEncoder;
-    type SyncPoint: Clone + Debug;
+    type SyncPoint: Clone + Debug + Default;
 
     fn create_command_encoder(&self, desc: super::CommandEncoderDesc) -> Self::CommandEncoder;
     fn destroy_command_encoder(&self, encoder: &mut Self::CommandEncoder);
-    fn submit(&self, encoder: &mut Self::CommandEncoder) -> Self::SyncPoint;
+    fn submit(
+        &self,
+        encoder: &mut Self::CommandEncoder,
+        after: &[Self::SyncPoint],
+    ) -> Self::SyncPoint;
     fn wait_for(&self, sp: &Self::SyncPoint, timeout_ms: u32) -> Result<bool, super::DeviceError>;
 }